[feat]: [PL-24523]: LDAP Password Secret Reference Support

+++ 400-rest/src/main/java/software/wings/beans/sso/LdapSettings.java
@@ -109,13 +109,14 @@ public class LdapSettings extends SSOSettings implements ExecutionCapabilityDema
 
   public EncryptedDataDetail getEncryptedDataDetails(SecretManager secretManager) {
     return secretManager
-        .encryptedDataDetails(
-            accountId, LdapConstants.BIND_PASSWORD_KEY, connectionSettings.getEncryptedBindPassword(), null)
+        .encryptedDataDetails(accountId, LdapConstants.BIND_PASSWORD_KEY, connectionSettings.fetchPassword(), null)
         .get();
   }
 
-  public void encryptFields(SecretManager secretManager) {
-    if (!connectionSettings.getBindPassword().equals(LdapConstants.MASKED_STRING)) {
+  public void encryptLdapInlineSecret(SecretManager secretManager) {
+    if (isNotEmpty(connectionSettings.getBindPassword())
+        && !connectionSettings.getBindPassword().equals(LdapConstants.MASKED_STRING)) {
+      connectionSettings.setPasswordType(LdapConnectionSettings.INLINE_SECRET);
       String oldEncryptedBindPassword = connectionSettings.getEncryptedBindPassword();
       if (isNotEmpty(oldEncryptedBindPassword)) {
         secretManager.deleteSecret(accountId, oldEncryptedBindPassword, new HashMap<>(), false);
@@ -128,6 +129,8 @@ public class LdapSettings extends SSOSettings implements ExecutionCapabilityDema
                                   .build();
       String encryptedBindPassword = secretManager.saveSecretText(accountId, secretText, false);
       connectionSettings.setEncryptedBindPassword(encryptedBindPassword);
+      connectionSettings.setBindPassword(LdapConstants.MASKED_STRING);
+    } else {
       connectionSettings.setBindPassword(LdapConstants.MASKED_STRING);
     }
   }
